tcbuild: Add udev rule to set iio permission
The application is executed as user torizon. Since the app need to be
able to write to sysfs DAC entries it is the straight forward solution
to set the affected file group to torizon and permission to 774.

It was necessary to also give the execution permission but it's not
clear why. Just set it to 774.